% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\name{cd}
\alias{cd}
\title{Coordinate Descent for SLOPE}
\usage{
cd(x, y, lambda, maxit = 100, opt = -Inf, opt_tol = 0.001)
}
\arguments{
\item{x}{design matrix}

\item{y}{response vector}

\item{lambda}{regularization}

\item{maxit}{max iterations}

\item{opt}{optimal value}

\item{opt_tol}{relative tolerance for optimal value}
}
\description{
Coordinate Descent for SLOPE
}
