% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdslope-package.R
\docType{package}
\name{cdslope-package}
\alias{cdslope}
\alias{cdslope-package}
\title{cdslope: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
