% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fista.R
\name{fista}
\alias{fista}
\title{Solve SLOPE with FISTA}
\usage{
fista(x, y, lambda, maxit = 100, opt = -Inf, opt_tol = 0.001)
}
\arguments{
\item{x}{design matrix}

\item{y}{response vector}

\item{lambda}{regularization}

\item{maxit}{max iterations}

\item{opt}{optimal value}

\item{opt_tol}{relative tolerance for optimal value}
}
\value{
A list of time, beta, and loss
}
\description{
Solve SLOPE with FISTA
}
